FinalProject/web/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
minio==7.1.17
mysql-connector-python